set annotations from the compiler api
