+++ libs/remix-solidity/src/compiler/compiler-abstract.ts
@@ -1,5 +1,5 @@
 'use strict'
-import txHelper from './txHelper'
+import helper from './helper'
 
 export class CompilerAbstract {
   languageversion: any
@@ -18,11 +18,11 @@ export class CompilerAbstract {
     }
 
     getContract (name) {
-      return txHelper.getContract(name, this.data.contracts)
+      return helper.getContract(name, this.data.contracts)
     }
 
     visitContracts (calllback) {
-      return txHelper.visitContracts(this.data.contracts, calllback)
+      return helper.visitContracts(this.data.contracts, calllback)
     }
 
     getData () {
