+++ libs/remix-ui/renderer/src/lib/renderer.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react' //eslint-disable-line
+import { helper } from '@remix-project/remix-solidity'
 import './renderer.css'
 interface RendererProps {
   message: any;
@@ -29,20 +30,11 @@ export const Renderer = ({ message, opt = {}, plugin }: RendererProps) => {
     // ^ e.g:
     // browser/gm.sol: Warning: Source file does not specify required compiler version! Consider adding "pragma solidity ^0.6.12
     // https://github.com/OpenZeppelin/openzeppelin-contracts/blob/release-v3.2.0/contracts/introspection/IERC1820Registry.sol:3:1: ParserError: Source file requires different compiler version (current compiler is 0.7.4+commit.3f05b770.Emscripten.clang) - note that nightly builds are considered to be strictly less than the released version
-    const positionDetails = getPositionDetails(text)
+    const positionDetails = helper.getPositionDetails(text)
 
     opt.errLine = positionDetails.errLine
     opt.errCol = positionDetails.errCol
     opt.errFile = positionDetails.errFile ? (positionDetails.errFile as string).trim() : ''
-
-    if (!opt.noAnnotations && opt.errFile && opt.errFile !== '') {
-      addAnnotation(opt.errFile, {
-        row: opt.errLine,
-        column: opt.errCol,
-        text: text,
-        type: opt.type
-      })
-    }
 
     setMessageText(text)
     setEditorOptions(opt)
@@ -50,30 +42,7 @@ export const Renderer = ({ message, opt = {}, plugin }: RendererProps) => {
     setClassList(opt.type === 'error' ? 'alert alert-danger' : 'alert alert-warning')
   }, [message, opt])
 
-  const getPositionDetails = (msg: string) => {
-    const result = { } as Record<string, number | string>
-
-    // To handle some compiler warning without location like SPDX license warning etc
-    if (!msg.includes(':')) return { errLine: -1, errCol: -1, errFile: '' }
-
-    if (msg.includes('-->')) msg = msg.split('-->')[1].trim()
-
-    // extract line / column
-    let pos = msg.match(/^(.*?):([0-9]*?):([0-9]*?)?/)
-    result.errLine = pos ? parseInt(pos[2]) - 1 : -1
-    result.errCol = pos ? parseInt(pos[3]) : -1
-
-    // extract file
-    pos = msg.match(/^(https:.*?|http:.*?|.*?):/)
-    result.errFile = pos ? pos[1] : msg
-    return result
-  }
-
-  const addAnnotation = async (file, error) => {
-    if (file === await plugin.call('config', 'getAppParameter', 'currentFile')) {
-      await plugin.call('editor', 'addAnnotation', error, file)
-    }
-  }
+  
 
   const handleErrorClick = (opt) => {
     if (opt.click) {
